combine all resource model tests into single file and convert openmeteo historical to parameterized
